packages/cortex-cli/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cortex-cli-agent"
version = "0.1.1"
description = "Personal Agent Infrastructure Stack — one install, infinite agents."
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.scripts]
cortex = "cortex_cli.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["cortex_cli"]

[project.optional-dependencies]
all = [
    "forge-agent",
    "a2a-hub",
    "vibe-replay",
    "agent-dispatcher",
]

[tool.uv.sources]
forge-agent = { workspace = true }
a2a-hub = { workspace = true }
vibe-replay = { workspace = true }
agent-dispatcher = { workspace = true }
